refactoring: - mutex replaced with a spin lock for creating critical regions - better use of atomi memory orders - better CI script
