+++ tests/Tests.cpp
@@ -3,118 +3,18 @@
 
 #include <CppThreadPool/CppThreadPool.hxx>
 
-#include <random>
+#include "UniformSampler.h"
+
 #include <stdexcept>
-#include <variant>
+#include <vector>
 
-namespace CppThreadPool {
-class UniformSampler {
-public:
-  UniformSampler() : generator(std::random_device{}()), distribution(1, 20) {
-    resetSeed(2);
-  }
-
-  Priority sample() const {
-    return static_cast<Priority>(this->distribution(this->generator));
-  };
-
-  void resetSeed(const std::size_t &newSeed) {
-    this->generator.seed(static_cast<unsigned int>(newSeed));
-  }
-
-private:
-  mutable std::mt19937 generator;
-  mutable std::uniform_int_distribution<> distribution;
-};
-
-enum class PoolType { FIFO, LIFO, PRIORITY };
-
-using AnyPool = std::variant<std::unique_ptr<Fifo>, std::unique_ptr<Lifo>,
-                             std::unique_ptr<Prioritized>>;
-
-class PoolWrapper {
-public:
-  PoolWrapper(const std::size_t size, const PoolType type) {
-    switch (type) {
-    case PoolType::FIFO:
-      pool = std::make_unique<Fifo>(size);
-      break;
-    case PoolType::LIFO:
-      pool = std::make_unique<Lifo>(size);
-      break;
-    case PoolType::PRIORITY:
-      pool = std::make_unique<Prioritized>(size);
-      break;
-    }
-  }
-
-  std::future<void> push(const std::function<void()> &action) {
-    std::future<void> result;
-    struct Visitor {
-      const std::function<void()> &action;
-      std::future<void> &result;
-      UniformSampler &priority_sampler;
-
-      void operator()(std::unique_ptr<Fifo> &pool) const {
-        result = pool->push(action);
-      }
-      void operator()(std::unique_ptr<Lifo> &pool) const {
-        result = pool->push(action);
-      }
-      void operator()(std::unique_ptr<Prioritized> &pool) const {
-        Priority p = priority_sampler.sample();
-        result = pool->push(action, p);
-      }
-    } visitor{action, result, priority_sampler};
-    std::visit(visitor, pool);
-    return result;
-  }
-
-  void wait() {
-    struct Visitor {
-      void operator()(std::unique_ptr<Fifo> &pool) const { pool->wait(); }
-      void operator()(std::unique_ptr<Lifo> &pool) const { pool->wait(); }
-      void operator()(std::unique_ptr<Prioritized> &pool) const {
-        pool->wait();
-      }
-    } visitor;
-    std::visit(visitor, pool);
-  }
-
-  std::size_t size() const {
-    std::size_t result;
-    struct Visitor {
-      std::size_t &result;
-
-      void operator()(const std::unique_ptr<Fifo> &pool) const {
-        result = pool->size();
-      }
-      void operator()(const std::unique_ptr<Lifo> &pool) const {
-        result = pool->size();
-      }
-      void operator()(const std::unique_ptr<Prioritized> &pool) const {
-        result = pool->size();
-      }
-    } visitor{result};
-    std::visit(visitor, pool);
-    return result;
-  }
-
-private:
-  AnyPool pool;
-  UniformSampler priority_sampler;
-};
-} // namespace CppThreadPool
-
-#define GENERATE_POOL_TYPES                                                    \
-  GENERATE(CppThreadPool::PoolType::FIFO, CppThreadPool::PoolType::LIFO,       \
-           CppThreadPool::PoolType::PRIORITY)
+using namespace CppThreadPool;
 
 TEST_CASE("Build destroy multiple times a pool", "[ThreadPool]") {
   const std::size_t size = 4;
-  auto kind = GENERATE_POOL_TYPES;
+
   for (std::size_t k = 0; k < 5; ++k) {
-    CppThreadPool::PoolWrapper pool(size, kind);
+    Fifo pool(size);
     CHECK(size == pool.size());
   }
 }
@@ -129,22 +29,21 @@ public:
   ExceptionTest() : std::runtime_error("") {}
 };
 
-void failure() { throw ExceptionTest{}; }
+void throw_exc() { throw ExceptionTest{}; }
 } // namespace
 
 TEST_CASE("Single task completed notification", "[ThreadPool]") {
   auto threads = GENERATE(1, 2, 3, 4);
-  auto kind = GENERATE_POOL_TYPES;
-  CppThreadPool::PoolWrapper pool(threads, kind);
+  Fifo pool{static_cast<std::size_t>(threads)};
 
-  SECTION("Single push and single wait") {
+  SECTION("Single push and single wait, success") {
     auto fut = pool.push(wait<500>);
     fut.wait();
     CHECK_NOTHROW(fut.get());
   }
 
-  SECTION("Single push and single wait") {
-    auto fut = pool.push(failure);
+  SECTION("Single push and single wait, failure") {
+    auto fut = pool.push(throw_exc);
     fut.wait();
     CHECK_THROWS_AS(fut.get(), ExceptionTest);
   }
@@ -152,11 +51,10 @@ TEST_CASE("Single task completed notification", "[ThreadPool]") {
 
 TEST_CASE("Multiple tasks completed notification", "[ThreadPool]") {
   auto threads = GENERATE(2, 4);
-  auto kind = GENERATE_POOL_TYPES;
-  CppThreadPool::PoolWrapper pool(threads, kind);
+  Fifo pool{static_cast<std::size_t>(threads)};
 
-  SECTION("Single push and single wait") {
-    std::list<std::future<void>> waiters;
+  SECTION("Multiple push and single wait, success") {
+    std::vector<std::future<void>> waiters;
     for (std::size_t k = 0; k < threads; ++k) {
       waiters.emplace_back(pool.push(wait<500>));
     }
@@ -166,10 +64,10 @@ TEST_CASE("Multiple tasks completed notification", "[ThreadPool]") {
     }
   }
 
-  SECTION("Single push and single wait") {
-    std::list<std::future<void>> waiters;
+  SECTION("Multiple push and single wait, failure") {
+    std::vector<std::future<void>> waiters;
     for (std::size_t k = 0; k < threads; ++k) {
-      waiters.emplace_back(pool.push(failure));
+      waiters.emplace_back(pool.push(throw_exc));
     }
     for (auto &w : waiters) {
       w.wait();
@@ -180,11 +78,10 @@ TEST_CASE("Multiple tasks completed notification", "[ThreadPool]") {
 
 TEST_CASE("Multiple tasks and wait for all", "[ThreadPool]") {
   auto threads = GENERATE(2, 4);
-  auto kind = GENERATE_POOL_TYPES;
-  CppThreadPool::PoolWrapper pool(threads, kind);
+  Fifo pool{static_cast<std::size_t>(threads)};
 
   for (std::size_t k = 0; k < 5; ++k) {
-    std::list<std::future<void>> waiters;
+    std::vector<std::future<void>> waiters;
     for (std::size_t k = 0; k < threads * 3; ++k) {
       waiters.emplace_back(pool.push(wait<500>));
     }
@@ -196,10 +93,10 @@ TEST_CASE("Multiple tasks and wait for all", "[ThreadPool]") {
 }
 
 namespace {
-std::chrono::milliseconds measure_time(const std::function<void()> &action) {
+template <typename Pred> std::chrono::microseconds measure_time(Pred action) {
   auto tic = std::chrono::high_resolution_clock::now();
   action();
-  return std::chrono::duration_cast<std::chrono::milliseconds>(
+  return std::chrono::duration_cast<std::chrono::microseconds>(
       std::chrono::high_resolution_clock::now() - tic);
 }
 } // namespace
@@ -213,7 +110,7 @@ TEST_CASE("Efficiency check", "[ThreadPool]") {
     }
   });
 
-  CppThreadPool::Fifo pool(2);
+  Fifo pool(2);
   auto parallel = measure_time([&]() {
     for (std::size_t t = 0; t < tasks; ++t) {
       pool.push(wait<150>);
@@ -224,3 +121,5 @@ TEST_CASE("Efficiency check", "[ThreadPool]") {
   CHECK(static_cast<double>(parallel.count()) <
         static_cast<double>(0.7 * serial.count()));
 }
+
+// TODO check other pool types
